Rename monitoring tests and update to remove selenium
